update subnet command
