+++ go.mod
@@ -2,7 +2,6 @@ module github.com/unicornultrafoundation/subnet-node
 
 go 1.24.0
 
-toolchain go1.24.4
 
 require (
 	github.com/gogo/protobuf v1.3.2
