--- script/k8s/provider/deployment.yaml
@@ -1,121 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: subnet-provider
-  namespace: subnet-provider
-  labels:
-    app: subnet-provider
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: subnet-provider
-  template:
-    metadata:
-      labels:
-        app: subnet-provider
-      annotations:
-        prometheus.io/scrape: "true"
-        prometheus.io/port: "9090"
-        prometheus.io/path: "/metrics"
-    spec:
-      containers:
-      - name: subnet-provider
-        image: subnet-node:latest
-        imagePullPolicy: IfNotPresent
-        ports:
-        - containerPort: 8080
-          name: http
-          protocol: TCP
-        - containerPort: 8081
-          name: health
-          protocol: TCP
-        - containerPort: 9090
-          name: metrics
-          protocol: TCP
-        env:
-        - name: PROVIDER_NAME
-          value: "subnet-provider"
-        - name: PROVIDER_PORT
-          value: "8080"
-        - name: HEALTH_PORT
-          value: "8081"
-        - name: METRICS_PORT
-          value: "9090"
-        - name: CONFIG_FILE
-          value: "/etc/config/config.yaml"
-        - name: PRIVATE_KEY
-          valueFrom:
-            secretKeyRef:
-              name: subnet-provider-secret
-              key: private-key
-        - name: IDENTITY_PRIVKEY
-          valueFrom:
-            secretKeyRef:
-              name: subnet-provider-secret
-              key: identity-privkey
-        - name: DATABASE_URL
-          valueFrom:
-            secretKeyRef:
-              name: subnet-provider-secret
-              key: database-url
-        - name: REDIS_PASSWORD
-          valueFrom:
-            secretKeyRef:
-              name: subnet-provider-secret
-              key: redis-password
-        resources:
-          requests:
-            cpu: 100m
-            memory: 128Mi
-          limits:
-            cpu: 500m
-            memory: 512Mi
-        volumeMounts:
-        - name: config-volume
-          mountPath: /etc/config
-          readOnly: true
-        - name: logs-volume
-          mountPath: /var/log/subnet
-        livenessProbe:
-          httpGet:
-            path: /health
-            port: health
-          initialDelaySeconds: 30
-          periodSeconds: 10
-          timeoutSeconds: 5
-          failureThreshold: 3
-        readinessProbe:
-          httpGet:
-            path: /ready
-            port: health
-          initialDelaySeconds: 5
-          periodSeconds: 5
-          timeoutSeconds: 3
-          failureThreshold: 3
-        startupProbe:
-          httpGet:
-            path: /health
-            port: health
-          initialDelaySeconds: 10
-          periodSeconds: 5
-          timeoutSeconds: 3
-          failureThreshold: 30
-        securityContext:
-          runAsNonRoot: true
-          runAsUser: 1000
-          runAsGroup: 1000
-          allowPrivilegeEscalation: false
-          readOnlyRootFilesystem: true
-          capabilities:
-            drop:
-            - ALL
-      volumes:
-      - name: config-volume
-        configMap:
-          name: subnet-provider-config
-      - name: logs-volume
-        emptyDir: {}
-      restartPolicy: Always
-      securityContext:
-        fsGroup: 1000
