--- script/k8s/provider/ingress.yaml
@@ -1,65 +1,0 @@
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: subnet-provider-ingress
-  namespace: subnet-provider
-  labels:
-    app: subnet-provider
-  annotations:
-    nginx.ingress.kubernetes.io/rewrite-target: /
-    nginx.ingress.kubernetes.io/ssl-redirect: "false"
-    nginx.ingress.kubernetes.io/rate-limit: "100"
-    nginx.ingress.kubernetes.io/rate-limit-window: "1m"
-    nginx.ingress.kubernetes.io/proxy-body-size: "10m"
-    nginx.ingress.kubernetes.io/proxy-connect-timeout: "30"
-    nginx.ingress.kubernetes.io/proxy-send-timeout: "30"
-    nginx.ingress.kubernetes.io/proxy-read-timeout: "30"
-spec:
-  ingressClassName: nginx
-  rules:
-  - host: provider.example.com  # Replace with your actual domain
-    http:
-      paths:
-      - path: /
-        pathType: Prefix
-        backend:
-          service:
-            name: subnet-provider
-            port:
-              number: 8080
-  - host: provider-api.example.com  # Replace with your actual API domain
-    http:
-      paths:
-      - path: /api
-        pathType: Prefix
-        backend:
-          service:
-            name: subnet-provider
-            port:
-              number: 8080
----
-# TLS configuration (uncomment and configure if using HTTPS)
-# apiVersion: networking.k8s.io/v1
-# kind: Ingress
-# metadata:
-#   name: subnet-provider-ingress-tls
-#   namespace: subnet-provider
-#   annotations:
-#     cert-manager.io/cluster-issuer: "letsencrypt-prod"
-# spec:
-#   tls:
-#   - hosts:
-#     - provider.example.com
-#     - provider-api.example.com
-#     secretName: subnet-provider-tls
-#   rules:
-#   - host: provider.example.com
-#     http:
-#       paths:
-#       - path: /
-#         pathType: Prefix
-#         backend:
-#           service:
-#             name: subnet-provider
-#             port:
-#               number: 8080
