--- script/k8s/provider/namespace.yaml
@@ -1,7 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: subnet-provider
-  labels:
-    name: subnet-provider
-    app: subnet-provider
