--- script/k8s/provider/secret.yaml
@@ -1,24 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: subnet-provider-secret
-  namespace: subnet-provider
-  labels:
-    app: subnet-provider
-type: Opaque
-data:
-  # Private key for blockchain account (base64 encoded)
-  private-key: eW91ci1wcml2YXRlLWtleS1oZXJl
-  # Identity private key for libp2p (base64 encoded)
-  identity-privkey: eW91ci1pZGVudGl0eS1wcml2a2V5LWhlcmU=
-  # Database URL - replace with your actual database URL (base64 encoded)
-  database-url: cG9zdGdyZXNxbDovL3Byb3ZpZGVyOnBhc3N3b3JkQHBvc3RncmVzOjU0MzIvc3VibmV0
-  # Redis password - replace with your actual Redis password (base64 encoded)
-  redis-password: cmVkaXMtcGFzc3dvcmQ=
-stringData:
-  # These will be automatically base64 encoded by Kubernetes
-  # You can use these for easier editing
-  # private-key: "your-private-key-here"
-  # identity-privkey: "your-identity-private-key-here"
-  # database-url: "postgresql://provider:password@postgres:5432/subnet"
-  # redis-password: "redis-password"
