--- script/k8s/provider/service.yaml
@@ -1,28 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: subnet-provider
-  namespace: subnet-provider
-  labels:
-    app: subnet-provider
-  annotations:
-    prometheus.io/scrape: "true"
-    prometheus.io/port: "9090"
-    prometheus.io/path: "/metrics"
-spec:
-  selector:
-    app: subnet-provider
-  ports:
-    - name: http
-      port: 8080
-      targetPort: 8080
-      protocol: TCP
-    - name: health
-      port: 8081
-      targetPort: 8081
-      protocol: TCP
-    - name: metrics
-      port: 9090
-      targetPort: 9090
-      protocol: TCP
-  type: ClusterIP
